fixed error where upgrade cards gets deleted from the board and where player is duplicated

+++ src/main/java/dk/dtu/compute/se/pisd/RoboSpring/RestFulAPI/Controller/PlayerController.java
@@ -25,7 +25,7 @@ public class PlayerController
     }
 
     @RequestMapping("players/leave")
-    public boolean leaveGame(Long gameID, int playerID)
+    public boolean leaveGame(Long gameID, Long playerID)
     {
         Player player = playerRepository.findPlayerByGameIDAndPlayerID(gameID, playerID);
         if (player == null)
